# docs/openapi-object-detection.yaml
openapi: 3.0.3
info:
  title: FlowGuard Object Detection & SecurePi Integration API
  description: >
    JSON REST endpoints for Camera Inventory, Monitoring Zones (Detection Setup),
    Detection Alerts, and the SecurePi edge-ingest route. Derived from
    server/routes/*.js and ai-service/main.py as implemented on
    branch feature/object-detection-space (2026-07-11).
    MJPEG endpoints (SecurePi /video_feed, AI Service /api/yolo/stream) are
    intentionally excluded — they are not JSON REST responses.
  version: "1.0.0"
servers:
  - url: http://localhost:5001
    description: FlowGuard Node.js backend (dev)
  - url: http://localhost:8501
    description: FlowGuard AI Service (FastAPI, dev) - reached by the frontend via the Vite "/ai" proxy
...
components:
  securitySchemes:
    bearerJwt:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: FlowGuard user JWT (FM/Staff/Tenant), verified against APP_SECRET.
    serviceKey:
      type: apiKey
      in: header
      name: x-service-key
      description: Shared secret used by the Python AI engine (AI_SERVICE_KEY).
    edgeIngestToken:
      type: http
      scheme: bearer
      description: Static device token used by the SecurePi edge node (EDGE_INGEST_TOKEN).
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
      example:
        error: "zone_name and camera_location are required."
    AuthError:
      type: object
      properties:
        message:
          type: string
      example:
        message: "Access Denied. No security token provided."
    MonitoringZone:
      type: object
      properties:
        id: { type: integer }
        zone_name: { type: string }
        location: { type: string }
        time_threshold: { type: integer, description: "Legacy threshold in minutes" }
        monitored_classes:
          type: array
          items: { type: string }
        density_threshold: { type: integer, nullable: true }
        unattended_threshold_seconds: { type: integer, nullable: true }
        alert_cooldown_seconds: { type: integer, nullable: true }
        severity: { type: string, enum: [Low, Medium, High, Critical] }
        assigned_team: { type: string, nullable: true }
        detection_enabled: { type: boolean }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    MonitoringZoneInput:
      type: object
      required: [zone_name, location, time_threshold]
      properties:
        zone_name: { type: string }
        location: { type: string }
        time_threshold: { type: number, minimum: 1 }
        density_threshold: { type: number, nullable: true }
        unattended_threshold_seconds: { type: number, nullable: true }
        alert_cooldown_seconds: { type: number, nullable: true }
        severity: { type: string, enum: [Low, Medium, High, Critical] }
        assigned_team: { type: string, nullable: true }
        detection_enabled: { type: boolean }
        monitored_classes:
          oneOf:
            - type: array
              items: { type: string }
            - type: string
              description: comma-separated class list
...
paths:
  /api/zones:
    get:
      tags: [Zones]
      summary: List monitoring zones (Detection Setup)
      security: [{ bearerJwt: [] }]
      responses:
        '200':
          description: List of monitoring zones
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/MonitoringZone' }
        '401': { description: Missing/invalid token, content: { application/json: { schema: { $ref: '#/components/schemas/AuthError' } } } }
        '403': { description: Role is not FM/Staff, content: { application/json: { schema: { $ref: '#/components/schemas/AuthError' } } } }
        '500': { description: Server error, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    post:
      tags: [Zones]
      summary: Create a monitoring zone
      security: [{ bearerJwt: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/MonitoringZoneInput' }
      responses:
        '201':
          description: Created zone
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MonitoringZone' }
        '400': { description: Validation error, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '401': { description: Missing/invalid token }
        '403': { description: Role is not FM }
        '500': { description: Server error }
  /api/zones/{id}:
    put:
      tags: [Zones]
      summary: Update a monitoring zone
      security: [{ bearerJwt: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/MonitoringZoneInput' }
      responses:
        '200':
          description: Updated zone
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MonitoringZone' }
        '400': { description: Validation error }
        '401': { description: Missing/invalid token }
        '403': { description: Role is not FM }
        '404': { description: Zone not found (empty body) }
        '500': { description: Server error }
    delete:
      tags: [Zones]
      summary: Delete a monitoring zone
      security: [{ bearerJwt: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200': { description: Deleted (empty body) }
        '401': { description: Missing/invalid token }
        '403': { description: Role is not FM }
        '404': { description: Zone not found (empty body) }
        '500': { description: Server error }
...